requirements.txt
confluent-kafka==0.11.6
kafka-python
boto3==1.9.75
SQLAlchemy==1.2.19
pyhocon==0.3.38
psycopg2==2.7.6.1